# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObjectPage": {
        "additionalProperties": false,
        "properties": {
          "collectedAt": {
            "title": "Collectedat",
            "type": "string"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "objectType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objecttype"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "rows": {
            "items": {
              "items": {},
              "type": "array"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "owner",
          "offset",
          "limit",
          "columns",
          "rows",
          "hasMore",
          "collectedAt"
        ],
        "title": "ObjectPage",
        "type": "object"
      },
...
      "PanelResult": {
        "additionalProperties": false,
        "description": "A diagnostic panel plus the state it was collected in.\n\n``available`` is false when the data could not be collected. It is never\npresented as an empty, healthy panel.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "collectedAt": {
            "title": "Collectedat",
            "type": "string"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "operationId": {
            "title": "Operationid",
            "type": "string"
          },
          "rows": {
            "items": {
              "items": {},
              "type": "array"
            },
            "title": "Rows",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "operationId",
          "title",
          "available",
          "collectedAt"
        ],
        "title": "PanelResult",
        "type": "object"
      },
      "ProfileIn": {
        "additionalProperties": false,
        "properties": {
          "defaultSchema": {
            "default": "",
            "title": "Defaultschema",
            "type": "string"
          },
          "environment": {
            "default": "development",
            "title": "Environment",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "mutatingRunbooksEnabled": {
            "default": false,
            "title": "Mutatingrunbooksenabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "port": {
            "default": 1521,
            "title": "Port",
            "type": "integer"
          },
          "protocol": {
            "default": "tcp",
            "title": "Protocol",
            "type": "string"
          },
          "secretReference": {
            "title": "Secretreference",
            "type": "string"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "walletDir": {
            "default": "",
            "title": "Walletdir",
            "type": "string"
          },
          "worksheetsEnabled": {
            "default": false,
            "title": "Worksheetsenabled",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "host",
          "serviceName",
          "username",
          "secretReference"
        ],
        "title": "ProfileIn",
        "type": "object"
      },
...
      "TargetView": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityView"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "defaultSchema": {
            "default": "",
            "title": "Defaultschema",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity"
          },
          "identityCheckedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identitycheckedat"
          },
          "mutatingRunbooksEnabled": {
            "title": "Mutatingrunbooksenabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "worksheetsEnabled": {
            "title": "Worksheetsenabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "environment",
          "host",
          "port",
          "serviceName",
          "username",
          "worksheetsEnabled",
          "mutatingRunbooksEnabled"
        ],
        "title": "TargetView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OracleDBHarness is the shared execution, policy and copilot layer for Oracle work.\n\nEvery operation carries an operation ID, an actor, a target, typed parameters,\ncapability requirements, a risk class and limits, and produces an execution record\nwith a structured result and verification evidence.\n\nNotes that apply to the whole API:\n\n* Free-form SQL runs only on targets where an administrator enabled worksheets. A\n  keyword filter is not a read-only boundary, so the Oracle account is the boundary.\n* Production targets are observation only in this release.\n* Applying a copilot proposal changes an editor buffer. It never runs, compiles or\n  commits anything.\n",
    "title": "OracleDBHarness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/targets": {
      "post": {
        "operationId": "create_target_api_v1_admin_targets_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Target Api V1 Admin Targets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Target",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/targets": {
      "get": {
        "description": "Only targets the caller actually holds a grant on.",
        "operationId": "list_targets_api_v1_targets_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TargetView"
                  },
                  "title": "Response List Targets Api V1 Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Targets",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}": {
      "get": {
        "operationId": "get_target_api_v1_targets__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Target",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}/dba/overview": {
      "get": {
        "operationId": "dba_overview_api_v1_targets__profile_id__dba_overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dba Overview Api V1 Targets  Profile Id  Dba Overview Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dba Overview",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}/objects": {
      "get": {
        "description": "One page of objects. Large lists are paged, never returned whole.",
        "operationId": "list_objects_api_v1_targets__profile_id__objects_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "objectType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Objecttype"
            }
          },
          {
            "in": "query",
            "name": "nameFilter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namefilter"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Objects",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}/objects/detail": {
      "get": {
        "description": "Columns, keys, indexes, source, errors and dependencies for one object.",
        "operationId": "object_detail_api_v1_targets__profile_id__objects_detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "objectName",
            "required": true,
            "schema": {
              "title": "Objectname",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "objectType",
            "required": true,
            "schema": {
              "title": "Objecttype",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Object Detail Api V1 Targets  Profile Id  Objects Detail Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Object Detail",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}/schemas": {
      "get": {
        "operationId": "list_schemas_api_v1_targets__profile_id__schemas_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Schemas",
        "tags": [
          "targets"
        ]
      }
    },
    "/api/v1/targets/{profile_id}/test": {
      "post": {
        "description": "Connect, prove the target identity, and probe every capability.",
        "operationId": "test_connection_api_v1_targets__profile_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "targets"
        ]
      }
    },
...
  }
}